Add Mutex to Locking Mechanism (#5)
Platform code has had a locking mechanism added such that calls to i2c buses are properly handled across threads and processes. Stratum, when run with ONLP support, calls platform functions in a multi-threaded way. If an operation in one thread configures the i2c multiplexers, while another thread is in the middle of a different request, the mux configuration is essentially unknowable.

This adds mutex locking to the recently added file locking system.